requirements.txt
discord.py==2.3.2
yt-dlp==2023.11.16
python-dotenv==1.0.0
Flask==2.3.3
gunicorn==21.2.0
nest_asyncio==1.5.8 